requirements.txt
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
exceptiongroup==1.3.0
iniconfig==2.1.0
packaging==25.0
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pluggy==1.6.0
pycparser==2.23
Pygments==2.19.2
pypdfium2==4.30.0
pytesseract==0.3.13
pytest==8.4.2
tomli==2.3.0
typing_extensions==4.15.0
